Updated output and main

+++ io/output.cpp
@@ -14,7 +14,7 @@ void output_csv(lotka_volterra::Simulation const& simulation,
   file << "t,x,y,H\n";
   for (std::size_t i = 0; i < simulation.steps(); ++i) {
     lotka_volterra::State const& state = simulation.stateAt(i);
-    file << i*simulation.dt() << "," << state.x << "," << state.y << "," << state.H << "\n";
+    file << static_cast<double>(i)*simulation.dt() << "," << state.x << "," << state.y << "," << state.H << "\n";
   }
 }
 } // namespace io
